Fix CSVWSourceIterator
in `boostrap()`: don't throw IllegalStateExcsption when `this.next` is NULL,
because that just means the next line is empty, causing the `hasNext()` method
to return false, which is desired behaviour.
Throwing this exception likely causes a consumer to crash because it doesn't
expect an exception upon reading an empty line.