chore: change signature of Access:getInputStream
It is unnecessary for getInputStream to declare all the different
exceptions the different instances of Access can throw. Instead, declare
general Exception and have the instances declare their own specific
exceptions.

A side effect of this change is that constructors of iterators and
streams will *have* to declare Exception to be caught, since they use the Access interface.

See
https://coderanch.com/t/399874/java/java/Methods-throwing-Exception-Interface
for further information regarding this side effect.