Adding support to additional relational databases.
Support for H2, Denodo Platform, AWS Athena and Snowflake databases. In combination with this support, RDBAccess has been extended so it can support fully JDBC connection properties and propagate them to the driver. This is required for AWS Athena, as additional connection properties are required for a successful connection to this DB.

Additionally, a change to RDBAccess.normalizeData has been done to avoid a NPE when the input data string is null.